Refactored all Interview Card components into their own directory and tested
