+++ grr-ui/src/templates/Homepage/interviewList.js
@@ -2,10 +2,9 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import PresenceVRNavBar from "../PresenceVRNavBar/PresenceVRNavBar"
 import { Button, Header, Card, Grid, Modal, List, Input, Image, Divider } from 'semantic-ui-react';
-import ConsumeInterview from './consumeInterview';
-import Interview from './interview'
-import EditInterview from './editInterview';
-import CancelInterview from './cancelInterview'
+import ConsumeInterview from '../../utils/consumeInterview';
+import Interview from '../InterviewCard/interview'
+import InterviewForm from '../InterviewCard/InterviewForm';
 import _ from 'lodash';
